content/canvas/test/webgl-mochitest/test_webgl_request_context.html

<!DOCTYPE HTML>
<title>WebGL test: 'webgl' and 'experimental-webgl' context requests succeed,
    'moz-webgl' context requests fail.</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<script src="driver-info.js"></script>
<body>
<canvas id="c1"></canvas>
<canvas id="c2"></canvas>
<script>

var testFunc = ok;

function testContextRetrieval(canvasId, creationId, shouldSucceed) {
  var canvas = document.getElementById(canvasId);
  ok(canvas, 'Invalid `canvasId`: ' + canvasId);

  var createdGL = canvas.getContext(creationId);
  if (shouldSucceed) {
    testFunc(createdGL, 'Request for \'' + creationId + '\' should succeed.');
  } else {
    ok(!createdGL, 'Request for \'' + creationId + '\' should fail.');
  }
}

SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv({'set': [
  ['webgl.force-enabled', true]
]}, function() {
  testContextRetrieval('c1', 'experimental-webgl', true);
  testContextRetrieval('c2', 'moz-webgl', false);
  SimpleTest.finish();
});

</script>
